fix ci

+++ pkg/executor/join/hash_join_stats.go
@@ -181,7 +181,7 @@ func (e *hashJoinRuntimeStatsV2) String() string {
 		buf.WriteString("build_hash_table:{total:")
 		buf.WriteString(execdetails.FormatDuration(time.Duration(e.fetchAndBuildHashTable)))
 		buf.WriteString(", fetch:")
-		buf.WriteString(execdetails.FormatDuration(time.Duration(int64(e.fetchAndBuildHashTable) - e.maxBuildHashTable - e.maxPartitionData)))
+		buf.WriteString(execdetails.FormatDuration(time.Duration(e.fetchAndBuildHashTable - e.maxBuildHashTable - e.maxPartitionData)))
 		buf.WriteString(", build:")
 		buf.WriteString(execdetails.FormatDuration(time.Duration(e.maxBuildHashTable + e.maxPartitionData)))
 		buf.WriteString("}")
